pydantic-example/pyproject.toml
[project]
name = "pydantic-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.0.14",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
